minimize padding for title, change how save data are store, empty tab list will not be saved

+++ dev/scripts/core/TabList/Item/DeleteButton/Init/addEventListeners.js
@@ -2,6 +2,10 @@ export function addEventListeners(deleteButton) {
   //mousedown happens before click (click is after blur, which hides the button away)
   deleteButton.addEventListener("mousedown", (e) => {
     e.preventDefault();
+
+    //0 = left click
+    if (e.button !== 0) return;
+
     focusOnNextItem(deleteButton);
     deleteButton.deleteItem();
   });
